+++ packages/jest-runtime/src/index.ts
@@ -5,9 +5,19 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import {URL, fileURLToPath} from 'url';
+import {URL, fileURLToPath, pathToFileURL} from 'url';
 import * as path from 'path';
-import {Script, compileFunction} from 'vm';
+import {
+  Script,
+  // @ts-ignore: experimental, not added to the types
+  SourceTextModule,
+  // @ts-ignore: experimental, not added to the types
+  SyntheticModule,
+  Context as VMContext,
+  // @ts-ignore: experimental, not added to the types
+  Module as VMModule,
+  compileFunction,
+} from 'vm';
 import * as nativeModule from 'module';
 import type {Config, Global} from '@jest/types';
 import type {
@@ -106,6 +116,8 @@ const EVAL_RESULT_VARIABLE = 'Object.<anonymous>';
 
 type RunScriptEvalResult = {[EVAL_RESULT_VARIABLE]: ModuleWrapper};
 
+const runtimeSupportsVmModules = typeof SourceTextModule === 'function';
+
 /* eslint-disable-next-line no-redeclare */
 class Runtime {
   private _cacheFS: CacheFS;
@@ -125,6 +137,7 @@ class Runtime {
   private _moduleMocker: typeof jestMock;
   private _isolatedModuleRegistry: ModuleRegistry | null;
   private _moduleRegistry: ModuleRegistry;
+  private _esmoduleRegistry: Map<string, VMModule>;
   private _needsCoverageMapped: Set<string>;
   private _resolver: Resolver;
   private _shouldAutoMock: boolean;
@@ -169,6 +182,7 @@ class Runtime {
     this._isolatedModuleRegistry = null;
     this._isolatedMockRegistry = null;
     this._moduleRegistry = new Map();
+    this._esmoduleRegistry = new Map();
     this._needsCoverageMapped = new Set();
     this._resolver = resolver;
     this._scriptTransformer = new ScriptTransformer(config);
@@ -300,6 +314,89 @@ class Runtime {
 
   static getCLIOptions(): typeof cliOptions {
     return cliOptions;
+  }
+
+  // unstable as it should be replaced by https://github.com/nodejs/modules/issues/393, and we don't want people to use it
+  unstable_shouldLoadAsEsm = Resolver.unstable_shouldLoadAsEsm;
+
+  private async loadEsmModule(
+    modulePath: Config.Path,
+    query = '',
+  ): Promise<VMModule> {
+    if (modulePath === '@jest/globals') {
+      // TODO: create a Synthetic Module for this. Will need to create a `jest` object without a `LocalModuleRequire`
+      throw new Error(
+        'Importing `@jest/globals` is not supported from ESM yet',
+      );
+    }
+
+    const cacheKey = modulePath + query;
+
+    if (!this._esmoduleRegistry.has(cacheKey)) {
+      invariant(
+        typeof this._environment.getVmContext === 'function',
+        'ES Modules are only supported if your test environment has the `getVmContext` function',
+      );
+
+      const context = this._environment.getVmContext();
+
+      invariant(context);
+
+      if (this._resolver.isCoreModule(modulePath)) {
+        const core = await this._importCoreModule(modulePath, context);
+        this._esmoduleRegistry.set(cacheKey, core);
+        return core;
+      }
+
+      const transformedFile = this.transformFile(modulePath, {
+        isInternalModule: false,
+        supportsDynamicImport: true,
+        supportsStaticESM: true,
+      });
+
+      const module = new SourceTextModule(transformedFile.code, {
+        context,
+        identifier: modulePath,
+        importModuleDynamically: (
+          specifier: string,
+          referencingModule: VMModule,
+        ) =>
+          this.loadEsmModule(
+            this._resolveModule(referencingModule.identifier, specifier),
+          ),
+        initializeImportMeta(meta: ImportMeta) {
+          meta.url = pathToFileURL(modulePath).href;
+        },
+      });
+
+      this._esmoduleRegistry.set(cacheKey, module);
+    }
+
+    const module = this._esmoduleRegistry.get(cacheKey);
+
+    invariant(module);
+
+    return module;
+  }
+
+  async unstable_importModule(
+    from: Config.Path,
+    moduleName?: string,
+  ): Promise<void> {
+    invariant(
+      runtimeSupportsVmModules,
+      'You need to run with a version of node that supports ES Modules in the VM API.',
+    );
+
+    const modulePath = this._resolveModule(from, moduleName);
+
+    const module = await this.loadEsmModule(modulePath);
+    await module.link((specifier: string, referencingModule: VMModule) =>
+      this.loadEsmModule(
+        this._resolveModule(referencingModule.identifier, specifier),
+      ),
+    );
+    await module.evaluate();
   }
 
   requireModule<T = unknown>(
@@ -577,6 +674,7 @@ class Runtime {
     this._isolatedMockRegistry = null;
     this._mockRegistry.clear();
     this._moduleRegistry.clear();
+    this._esmoduleRegistry.clear();
 
     if (this._environment) {
       if (this._environment.global) {
@@ -795,23 +893,8 @@ class Runtime {
     Object.defineProperty(localModule, 'require', {
       value: this._createRequireImplementation(localModule, options),
     });
-    const transformedFile = this._scriptTransformer.transform(
-      filename,
-      this._getFullTransformationOptions(options),
-      this._cacheFS[filename],
-    );
 
-    // we only care about non-internal modules
-    if (!options || !options.isInternalModule) {
-      this._fileTransforms.set(filename, transformedFile);
-    }
-
-    if (transformedFile.sourceMapPath) {
-      this._sourceMapRegistry[filename] = transformedFile.sourceMapPath;
-      if (transformedFile.mapCoverage) {
-        this._needsCoverageMapped.add(filename);
-      }
-    }
+    const transformedFile = this.transformFile(filename, options);
 
     let compiledFunction: ModuleWrapper | null = null;
 
@@ -907,6 +990,27 @@ class Runtime {
     this._currentlyExecutingModulePath = lastExecutingModulePath;
   }
 
+  private transformFile(filename: string, options?: InternalModuleOptions) {
+    const transformedFile = this._scriptTransformer.transform(
+      filename,
+      this._getFullTransformationOptions(options),
+      this._cacheFS[filename],
+    );
+
+    // we only care about non-internal modules
+    if (!options || !options.isInternalModule) {
+      this._fileTransforms.set(filename, transformedFile);
+    }
+
+    if (transformedFile.sourceMapPath) {
+      this._sourceMapRegistry[filename] = transformedFile.sourceMapPath;
+      if (transformedFile.mapCoverage) {
+        this._needsCoverageMapped.add(filename);
+      }
+    }
+    return transformedFile;
+  }
+
   private createScriptFromCode(scriptSource: string, filename: string) {
     try {
       return new Script(this.wrapCodeInModuleWrapper(scriptSource), {
@@ -930,6 +1034,24 @@ class Runtime {
     }
 
     return require(moduleName);
+  }
+
+  private _importCoreModule(moduleName: string, context: VMContext) {
+    const required = this._requireCoreModule(moduleName);
+
+    return new SyntheticModule(
+      ['default', ...Object.keys(required)],
+      function () {
+        // @ts-ignore: TS doesn't know what `this` is
+        this.setExport('default', required);
+        Object.entries(required).forEach(([key, value]) => {
+          // @ts-ignore: TS doesn't know what `this` is
+          this.setExport(key, value);
+        });
+      },
+      // should identifier be `node://${moduleName}`?
+      {context, identifier: moduleName},
+    );
   }
 
   private _getMockedNativeModule(): typeof nativeModule.Module {
@@ -966,13 +1088,20 @@ class Runtime {
     // should we implement the class ourselves?
     class Module extends nativeModule.Module {}
 
+    Object.entries(nativeModule.Module).forEach(([key, value]) => {
+      // @ts-ignore
+      Module[key] = value;
+    });
+
     Module.Module = Module;
 
     if ('createRequire' in nativeModule) {
       Module.createRequire = createRequire;
     }
     if ('createRequireFromPath' in nativeModule) {
-      Module.createRequireFromPath = (filename: string | URL) => {
+      Module.createRequireFromPath = function createRequireFromPath(
+        filename: string | URL,
+      ) {
         if (typeof filename !== 'string') {
           const error = new TypeError(
             `The argument 'filename' must be string. Received '${filename}'.${
@@ -989,7 +1118,7 @@ class Runtime {
       };
     }
     if ('syncBuiltinESMExports' in nativeModule) {
-      Module.syncBuiltinESMExports = () => {};
+      Module.syncBuiltinESMExports = function syncBuiltinESMExports() {};
     }
 
     this._moduleImplementation = Module;
@@ -1358,6 +1487,7 @@ class Runtime {
   private getGlobalsForFile(from: Config.Path): JestGlobalsValues {
     const jest = this.jestObjectCaches.get(from);
 
+    // This won't exist in ESM
     invariant(jest, 'There should always be a Jest object already');
 
     return {